Fix: Pop-up dialog for cancelation of the key upload appears on the wrong screen (EXPOSUREAPP-2915) (#1331)
* Removed submission cancellation dialog and replaced back handler with simple back navigation


* Readded submission cancellation warning to first fragment in new submission flow


* Switched back navigation to use findNavController().popBackStack()
