+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/fragment/SubmissionResultPositiveOtherWarningFragment.kt
@@ -4,7 +4,6 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.view.accessibility.AccessibilityEvent
-import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
@@ -35,14 +34,6 @@ class SubmissionResultPositiveOtherWarningFragment :
     private val binding: FragmentSubmissionPositiveOtherWarningBinding by viewBindingLazy()
     private lateinit var internalExposureNotificationPermissionHelper:
         InternalExposureNotificationPermissionHelper
-
-    // Overrides default back behaviour
-    private val backCallback: OnBackPressedCallback =
-        object : OnBackPressedCallback(true) {
-            override fun handleOnBackPressed() {
-                handleSubmissionCancellation()
-            }
-        }
 
     override fun onResume() {
         super.onResume()
@@ -100,7 +91,6 @@ class SubmissionResultPositiveOtherWarningFragment :
 
         internalExposureNotificationPermissionHelper =
             InternalExposureNotificationPermissionHelper(this, this)
-        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, backCallback)
 
         setButtonOnClickListener()
 
@@ -120,28 +110,8 @@ class SubmissionResultPositiveOtherWarningFragment :
             initiateWarningOthers()
         }
         binding.submissionPositiveOtherWarningHeader.headerButtonBack.buttonIcon.setOnClickListener {
-            handleSubmissionCancellation()
+            findNavController().popBackStack()
         }
-    }
-
-    /**
-     * Opens a Dialog that warns user
-     * when they're about to cancel the submission flow
-     * @see DialogHelper
-     * @see navigateToSubmissionResultFragment
-     */
-    fun handleSubmissionCancellation() {
-        DialogHelper.showDialog(
-            DialogHelper.DialogInstance(
-                requireActivity(),
-                R.string.submission_error_dialog_confirm_cancellation_title,
-                R.string.submission_error_dialog_confirm_cancellation_body,
-                R.string.submission_error_dialog_confirm_cancellation_button_positive,
-                R.string.submission_error_dialog_confirm_cancellation_button_negative,
-                true,
-                ::navigateToSubmissionResultFragment
-            )
-        )
     }
 
     private fun navigateToSubmissionResultFragment() =
